#103 whitelist living person details - on both migrate and search
